Split widget: computed table header
